stol to stoi 2

+++ include/grad_aff/rap/RapParser.h
@@ -180,7 +180,7 @@ namespace grad_aff::RapParser2
                 rapVal->subType = 1;
             }
             else {
-                rapVal->value = std::stol(dec);
+                rapVal->value = std::stoi(dec);
                 rapVal->type = 1;
                 rapVal->subType = 2;
             }
